server/microservice-users/requirements.txt
anyio==3.6.2
click==8.1.3
fastapi==0.89.0
h11==0.14.0
httptools==0.5.0
idna==3.4
pydantic==1.10.4
PyJWT==2.6.0
python-dotenv==0.21.0
PyYAML==6.0
sniffio==1.3.0
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a31
sqlmodel==0.0.8
starlette==0.22.0
typing_extensions==4.4.0
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
